Separate component to render user poaps
